+++ NEWS.md
@@ -35,3 +35,5 @@
 * Updated color mapping for "plotBarcodeHeamtmap" with alternative data transformations.
 
 * Updated "BarcodePairCorrelation.R" by computing barcode pair-wise correlation and mean proportion based on transformed data.
+
+* Updated ""plotSamplePairCorrelation.R" by computing sample pair-wise correlation based on transformed data.